# Speed up finished runs MongoDb query