Remove skip=0

See https://github.com/glinscott/fishtest/pull/437

Edit: this speeds up from 3.5s to 0.8s

Filtering 'deleted' in python speeds up from 0.8s to 0.2s