@ppigazzini 2019-11-26
@tomtor dev-server test before your last commit (two runs) https://github.com/glinscott/fishtest/pull/443/commits/cac2c52ed4cecc63706d4d212be38746815a3049

* master:
```
${VENV}/bin/python fishtest/fishtest/utils/test_queries.p                      y

Fetching unfinished runs ...
0.000162839889526s
Fetching machines ...
0.0795078277588s
Fetching finished runs ...
2.60631203651s
Requesting pgn ...
0.000993013381958s

${VENV}/bin/python fishtest/fishtest/utils/test_queries.p                      y

Fetching unfinished runs ...
0.000190019607544s
Fetching machines ...
0.073038816452s
Fetching finished runs ...
2.89345192909s
Requesting pgn ...
0.000911951065063s
```
* pr:
```
${VENV}/bin/python fishtest/fishtest/utils/test_queries.p                      y

Fetching unfinished runs ...
0.000173091888428s
Fetching machines ...
0.0736258029938s
Fetching finished runs ...
2.87277507782s
Requesting pgn ...
0.000942945480347s

$ ${VENV}/bin/python fishtest/fishtest/utils/test_queries.p                      y

Fetching unfinished runs ...
0.000170946121216s
Fetching machines ...
0.072921037674s
Fetching finished runs ...
2.79401803017s
Requesting pgn ...
0.000946998596191s
```