@tomtor 2019-11-26
@ppigazzini See ~/x/rebuild_queries on prod which I used for my timing...