@ppigazzini 2019-11-26
@tomtor dev-serve test **after** https://github.com/glinscott/fishtest/pull/443/commits/cac2c52ed4cecc63706d4d212be38746815a3049
```
$ ${VENV}/bin/python ${UPATH}/test_queries.py

Fetching unfinished runs ...
0.000162839889526s
Fetching machines ...
0.0793800354004s
Fetching finished runs ...
1.83845090866s
Requesting pgn ...
0.000683069229126s

$ ${VENV}/bin/python ${UPATH}/test_queries.py

Fetching unfinished runs ...
0.000181913375854s
Fetching machines ...
0.075838804245s
Fetching finished runs ...
1.90634894371s
Requesting pgn ...
0.000868082046509s
```