@ppigazzini 2019-11-26
@tomtor I viewed only now your post
> @ppigazzini See ~/x/rebuild_queries on prod which I used for my timing...