@ppigazzini 2019-11-26
@tomtor sorry for the confusion, after viewing your https://github.com/glinscott/fishtest/pull/443#issuecomment-558587099 I wanted to don't post my https://github.com/glinscott/fishtest/pull/443#issuecomment-558598854 and in automatic way I pushed the rightmost button. Unfortunately the button is "Close and comment" and not "Cancel" :(