@tomtor 2019-11-26
@ppigazzini I see you used `rebuild_indices.py`

So on prod the speedup is very large.