@ppigazzini 2019-11-26
@tomtor I'm not able to reproduce the speedup on the dev with `~/x/rebuild_indeces` (`~/x/rebuild_queries` not found).

*master:
```
$ ${VENV}/bin/python rebuild_indeces.py
reindex
t0=1574774084.04
t1=1574774084.18
t1-t0=0.138878107071
done...

$ ${VENV}/bin/python rebuild_indeces.py
reindex
t0 = 1574774122.86
t1 = 1574774123.02
t1-t0 = 0.157711029053
done...

$ ${VENV}/bin/python rebuild_indeces.py
reindex
t0 = 1574774191.32
t1 = 1574774191.46
t1-t0 = 0.139991044998
done...
```

*pr:
```
$ ${VENV}/bin/python rebuild_indeces.py
reindex
t0 = 1574774548.6
t1 = 1574774548.74
t1-t0 = 0.144069910049
done...
$ ${VENV}/bin/python rebuild_indeces.py
reindex
t0 = 1574774550.5
t1 = 1574774550.65
t1-t0 = 0.148473978043
done...
$ ${VENV}/bin/python rebuild_indeces.py
reindex
t0 = 1574774552.04
t1 = 1574774552.19
t1-t0 = 0.142319917679
done...
```