@xoto10 2019-11-26
Nice work!

I have made a new branch with some changes to test_queries.py that may give some useful info on this: [testqueries2](https://github.com/glinscott/fishtest/compare/master...xoto10:testqueries2)

It displays the explain() output for the old and new versions of the finished() query and tries using a hint().