@ppigazzini 2019-11-26
@tomtor this is on prod, I double checked that it's in sync w/ master (`git branch`, `git status` and inspected `rundb.py`) and that mongo has the usual indexes. I restarted mongod and fishtest.
I'm obviously missing something...

* master:
```
$ ${VENV}/bin/python rebuild_indices.py
reindex
t0 = 1574778108.66
t1 = 1574778108.82
t1-t0 = 0.156199932098
done...
$ ${VENV}/bin/python rebuild_indices.py
reindex
t0 = 1574778110.97
t1 = 1574778111.14
t1-t0 = 0.168743133545
done...
```

* pr (files copied by hand):
```
${VENV}/bin/python rebuild_indices.py
reindex
t0 = 1574778777.21
t1 = 1574778777.39
t1-t0 = 0.185088157654
done...
$ ${VENV}/bin/python rebuild_indices.py
reindex
t0 = 1574778779.22
t1 = 1574778779.43
t1-t0 = 0.21334695816
done...
```