@tomtor 2019-11-26
@ppigazzini rebuild_indices only does the new query from this PR (in 0.2s instead of eg 2.0s) and is not depending on the PR. If you re-add skip=0 and the test for `deleted` it will slow down again.