@tomtor 2019-11-26
t1-t0 = 0.167485952377         (PR)                                                                                                                     
    t1-t0 = 0.709129095078         (master)