@tomtor 2019-11-26
@ppigazzini For some reason the `skip=0` no longer makes a major difference, but 0.16s for the main page is very fast...