@tomtor 2019-11-26
I removed the `skip=0` rewrite and only left the removal of `deleted` from the query....