@ppigazzini 2019-11-26
@tomtor it's a bit paranoid but on dev, when testing for speedup, I comment out crontab, I run the scripts a couple of times to avoid cache effects, and I reinstall fishtest with `update_fishtest.py` to be sure to don't use leftovers (*.pyc)