@ppigazzini 2019-11-26
@tomtor I wrote this `test_pr443.py`, check if it's correct:
```python
from __future__ import print_function

import os
import sys
import time
import numpy
from pymongo import MongoClient, ASCENDING, DESCENDING

db_name='fishtest_new'
# MongoDB server is assumed to be on the same machine, if not user should use
# ssh with port forwarding to access the remote host.
conn = MongoClient(os.getenv('FISHTEST_HOST') or 'localhost')
db = conn[db_name]

def test_query(q):
  n=20
  t=[]
  t.append(time.time())
  for i in range(n):
    c = db['runs'].find(q, skip=0, limit=100, sort=[('last_updated', DESCENDING)])
    result = [list(c), c.count()]
    t.append(time.time())

  delta=numpy.array([j-i for i,j in zip(t[:-1], t[1:])])

  print("query =%s" % q)
  print("mean = %s" % numpy.mean(delta))
  print("stdev = %s" % numpy.std(delta))

ma = {'finished': True, 'deleted': {'$exists': False}}
pr ={'finished': True}
test_query(ma)
test_query(pr)

print("done...")
```

Result:
```
$ ${VENV}/bin/python test_pr443.py
query ={'deleted': {'$exists': False}, 'finished': True}
mean = 0.23638049364089966
std = 0.037089635227488704
query ={'finished': True}
mean = 0.19297879934310913
std = 0.029167710507801748
done...
```

Result w/ comment out `result = [list(c), c.count()]`:
```
$ ${VENV}/bin/python test_pr443.py
query ={'deleted': {'$exists': False}, 'finished': True}
mean = 5.32984733581543e-05
std = 4.1316809902932895e-05
query ={'finished': True}
mean = 3.159046173095703e-05
std = 5.517471610339806e-06
done...
```