@tomtor 2019-11-26
@ppigazzini Why don't you compare the PR to:

    ${VENV}/bin/python fishtest/fishtest/utils/test_queries.py

which took 2.6s in master? (first tests in this PR)