@ppigazzini 2019-11-26
@tomtor I compared the two queries from `rebuild_indices.py`, if I'm not wrong those are the queries swapped with this PR.
By the way swapping the queries order in `rebuild_indices.py` I got reversed results (and results very noisy), so I wrote that new python script. I don't know if the mongodb cache system voids the results.